Bump spew godep to make hash tests pass on go 1.6.
